Move ObjectFactories class to spi package. Move `newServiceFactory` class to ServiceProvision class.
